Make `schemaVersion` required, remove alpha note, combine files
