+++ src/handlers/shared/saleor-webhook.ts
@@ -1,4 +1,4 @@
-import { SaleorWebhookError } from "./process-saleor-webhook";
+import { AuthData } from "../../APL";
 
 export const WebhookErrorCodeMap: Record<SaleorWebhookError, number> = {
   OTHER: 500,
@@ -15,4 +15,41 @@ export const WebhookErrorCodeMap: Record<SaleorWebhookError, number> = {
   WRONG_METHOD: 405,
   CANT_BE_PARSED: 400,
   CONFIGURATION_ERROR: 500,
+};
+
+export type SaleorWebhookError =
+  | "OTHER"
+  | "MISSING_HOST_HEADER"
+  | "MISSING_DOMAIN_HEADER"
+  | "MISSING_API_URL_HEADER"
+  | "MISSING_EVENT_HEADER"
+  | "MISSING_PAYLOAD_HEADER"
+  | "MISSING_SIGNATURE_HEADER"
+  | "MISSING_REQUEST_BODY"
+  | "WRONG_EVENT"
+  | "NOT_REGISTERED"
+  | "SIGNATURE_VERIFICATION_FAILED"
+  | "WRONG_METHOD"
+  | "CANT_BE_PARSED"
+  | "CONFIGURATION_ERROR";
+
+export class WebhookError extends Error {
+  errorType: SaleorWebhookError = "OTHER";
+
+  constructor(message: string, errorType: SaleorWebhookError) {
+    super(message);
+    if (errorType) {
+      this.errorType = errorType;
+    }
+    Object.setPrototypeOf(this, WebhookError.prototype);
+  }
+}
+
+export type WebhookContext<TPayload> = {
+  baseUrl: string;
+  event: string;
+  payload: TPayload;
+  authData: AuthData;
+  /** Added in Saleor 3.15 */
+  schemaVersion: number;
 };
